chore: move crosscontroller tests into main package
Also relocates the mock watcher into package_test and delete the mock
file.
